[form-factors] Move ``KMPW2010`` parametrization to separate files
