+++ eos/form-factors/Makefile.am
@@ -30,6 +30,7 @@ libeosformfactors_la_SOURCES = \
 	parametric-bgl1997.hh parametric-bgl1997-impl.hh parametric-bgl1997.cc \
 	parametric-bmrvd2022.hh parametric-bmrvd2022-impl.hh parametric-bmrvd2022.cc \
 	parametric-bsz2015.hh parametric-bsz2015-impl.hh parametric-bsz2015.cc \
+	parametric-kmpw2010.hh parametric-kmpw2010-impl.hh parametric-kmpw2010.cc \
 	pi-lcdas.cc pi-lcdas.hh \
 	unitarity-bounds.cc unitarity-bounds.hh \
 	zero-recoil-sum-rule.cc zero-recoil-sum-rule.hh
@@ -53,6 +54,7 @@ include_eos_rarebdecays_HEADERS = \
 	parametric-bcl2008.hh \
 	parametric-bgl1997.hh \
 	parametric-bsz2015.hh \
+	parametric-kmpw2010.hh \
 	pi-lcdas.hh
 
 AM_TESTS_ENVIRONMENT = \
@@ -74,6 +76,7 @@ TESTS = \
 	parametric-bgl1997_TEST \
 	parametric-bmrvd2022_TEST \
 	parametric-bsz2015_TEST \
+	parametric-kmpw2010_TEST \
 	pi-lcdas_TEST \
 	unitarity-bounds_TEST
 LDADD = \
@@ -116,6 +119,8 @@ parametric_bgl1997_TEST_SOURCES = parametric-bgl1997_TEST.cc
 parametric_bsz2015_TEST_SOURCES = parametric-bsz2015_TEST.cc
 
 parametric_bmrvd2022_TEST_SOURCES = parametric-bmrvd2022_TEST.cc
+
+parametric_kmpw2010_TEST_SOURCES = parametric-kmpw2010_TEST.cc
 
 pi_lcdas_TEST_SOURCES = pi-lcdas_TEST.cc
