+++ eos/form-factors/mesonic-impl.hh
@@ -46,204 +46,6 @@ namespace eos
     // Process    = B -> K^*, B -> D^*, B -> rho, B_s -> phi etc.
     // Transition = B -> V or B -> P
 
-    /* Form Factors according to [KMPW2010] */
-    template <typename Transition_> class KMPW2010FormFactors;
-
-    template <> class KMPW2010FormFactors<PToV> :
-        public FormFactors<PToV>
-    {
-        private:
-            // fit parametrisation for P -> V according to [KMPW2010]
-            UsedParameter
-               _f0_V, _b1_V,
-               _f0_A0, _b1_A0, _f0_A1, _b1_A1, _f0_A2, _b1_A2,
-               _f0_T1, _b1_T1, _f0_T2, _b1_T2, _f0_T3, _b1_T3;
-
-            static const double _tau_p, _tau_m, _tau_0;
-            static const double _m_B, _m_Kstar, _m_Bs2_0m, _m_Bs2_1m, _m_Bs2_1p;
-
-            static double _calc_z(const double & s)
-            {
-                return (std::sqrt(_tau_p - s) - std::sqrt(_tau_p - _tau_0)) / (std::sqrt(_tau_p - s) + std::sqrt(_tau_p - _tau_0));
-            }
-
-            static double ff_KMPW(const double & s, const double & f0, const double & b1, const double & m2)
-            {
-                const double zs = _calc_z(s), z0 = _calc_z(0.0);
-
-                // cf. [KMPW2010], Eq. (8.8), p. 30
-                return f0 / (1.0 - s / m2) * (1.0 + b1 * (zs - z0 + 0.5 * (zs * zs - z0 * z0)));
-            }
-
-        // Remark: this is hardcoded B->K* FF's nuisance parameters (OK, because in [KMPW2010] only B->K^* calculated)
-        public:
-            KMPW2010FormFactors(const Parameters & p, const Options &) :
-                _f0_V(p["B->K^*::F^V(0)@KMPW2010"],   *this),   _b1_V(p["B->K^*::b^V_1@KMPW2010"],    *this),
-                _f0_A0(p["B->K^*::F^A0(0)@KMPW2010"], *this),   _b1_A0(p["B->K^*::b^A0_1@KMPW2010"],  *this),
-                _f0_A1(p["B->K^*::F^A1(0)@KMPW2010"], *this),   _b1_A1(p["B->K^*::b^A1_1@KMPW2010"],  *this),
-                _f0_A2(p["B->K^*::F^A2(0)@KMPW2010"], *this),   _b1_A2(p["B->K^*::b^A2_1@KMPW2010"],  *this),
-                _f0_T1(p["B->K^*::F^T1(0)@KMPW2010"], *this),   _b1_T1(p["B->K^*::b^T1_1@KMPW2010"],  *this),
-                _f0_T2(p["B->K^*::F^T2(0)@KMPW2010"], *this),   _b1_T2(p["B->K^*::b^T2_1@KMPW2010"],  *this),
-                _f0_T3(p["B->K^*::F^T3(0)@KMPW2010"], *this),   _b1_T3(p["B->K^*::b^T3_1@KMPW2010"],  *this)
-            {
-            }
-
-            static FormFactors<PToV> * make(const Parameters & parameters, const Options & options)
-            {
-                return new KMPW2010FormFactors(parameters, options);
-            }
-
-            virtual double v(const double & s) const
-            {
-                return ff_KMPW(s, _f0_V(), _b1_V(), _m_Bs2_1m);
-            }
-
-            virtual double a_0(const double & s) const
-            {
-                return ff_KMPW(s, _f0_A0(), _b1_A0(), _m_Bs2_0m);
-            }
-
-            virtual double a_1(const double & s) const
-            {
-                return ff_KMPW(s, _f0_A1(), _b1_A1(), _m_Bs2_1p);
-            }
-
-            virtual double a_2(const double & s) const
-            {
-                return ff_KMPW(s, _f0_A2(), _b1_A2(), _m_Bs2_1p);
-            }
-
-            virtual double a_12(const double & s) const
-            {
-                const double mB = BToKstar::mB, mB2 = mB * mB;
-                const double mV = BToKstar::mV, mV2 = mV * mV;
-                const double lambda = eos::lambda(mB2, mV2, s);
-
-                return ((mB + mV) * (mB + mV) * (mB2 - mV2 - s) * this->a_1(s)
-                    - lambda * this->a_2(s)) / (16.0 * mB * mV2 * (mB + mV));
-            }
-
-            virtual double t_1(const double & s) const
-            {
-                return ff_KMPW(s, _f0_T1(), _b1_T1(), _m_Bs2_1m);
-            }
-
-            virtual double t_2(const double & s) const
-            {
-                return ff_KMPW(s, _f0_T2(), _b1_T2(), _m_Bs2_1p);
-            }
-
-            virtual double t_3(const double & s) const
-            {
-                return ff_KMPW(s, _f0_T3(), _b1_T3(), _m_Bs2_1p);
-            }
-
-            virtual double t_23(const double & s) const
-            {
-                const double mB = BToKstar::mB, mB2 = mB * mB;
-                const double mV = BToKstar::mV, mV2 = mV * mV;
-                const double lambda = eos::lambda(mB2, mV2, s);
-
-                return ((mB2 - mV2) * (mB2 + 3.0 * mV2 - s) * this->t_2(s)
-                        - lambda * this->t_3(s)) / (8.0 * mB * mV2 * (mB - mV));
-            }
-
-            virtual double f_perp(const double &) const
-            {
-                return 0.;
-            }
-
-            virtual double f_para(const double &) const
-            {
-                return 0.;
-            }
-
-            virtual double f_long(const double &) const
-            {
-                return 0.;
-            }
-
-            virtual double f_perp_T(const double &) const
-            {
-                return 0.;
-            }
-
-            virtual double f_para_T(const double &) const
-            {
-                return 0.;
-            }
-
-            virtual double f_long_T(const double &) const
-            {
-                return 0.;
-            }
-
-            virtual double f_long_T_Normalized(const double &) const
-            {
-                return 0.;
-            }
-    };
-
-    template <> class KMPW2010FormFactors<PToP> :
-        public FormFactors<PToP>
-    {
-        private:
-            // fit parametrisation for P -> P according to [KMPW2010]
-            UsedParameter _b1_p, _b1_0, _b1_t;
-            UsedParameter _f0_p, _f0_t;
-            static const double _tau_p, _tau_m, _tau_0;
-            static const double _m_B, _m_K, _m_Bs2;
-
-            static double _calc_z(const double & s)
-            {
-                return (std::sqrt(_tau_p - s) - std::sqrt(_tau_p - _tau_0)) / (std::sqrt(_tau_p - s) + std::sqrt(_tau_p - _tau_0));
-            }
-
-        public:
-            KMPW2010FormFactors(const Parameters & p, const Options &) :
-                _b1_p(p["B->K::b^p_1@KMPW2010"], *this),
-                _b1_0(p["B->K::b^0_1@KMPW2010"], *this),
-                _b1_t(p["B->K::b^t_1@KMPW2010"], *this),
-                _f0_p(p["B->K::F^p(0)@KMPW2010"], *this),
-                _f0_t(p["B->K::F^t(0)@KMPW2010"], *this)
-            {
-            }
-
-            static FormFactors<PToP> * make(const Parameters & parameters, const Options & options)
-            {
-                return new KMPW2010FormFactors(parameters, options);
-            }
-
-            virtual double f_p(const double & s) const
-            {
-                // cf. [KMPW2010], Eq. (8.8), p. 30
-                const double zs = _calc_z(s), z0 = _calc_z(0.0);
-
-                return _f0_p() / (1 - s / _m_Bs2) * (1 + _b1_p() * (zs - z0 + 0.5 * (zs * zs - z0 * z0)));
-            }
-
-            virtual double f_0(const double & s) const
-            {
-                // cf. [KMPW2010], Eq. (8.8), p. 30
-                const double zs = _calc_z(s), z0 = _calc_z(0.0);
-
-                return _f0_p() * (1 + _b1_0() * (zs - z0 + 0.5 * (zs * zs - z0 * z0)));
-            }
-
-            virtual double f_t(const double & s) const
-            {
-                // cf. [KMPW2010], Eq. (8.8), p. 30
-                const double zs = _calc_z(s), z0 = _calc_z(0.0);
-
-                return _f0_t() / (1 - s / _m_Bs2) * (1 + _b1_t() * (zs - z0 + 0.5 * (zs * zs - z0 * z0)));
-            }
-
-            virtual double f_plus_T(const double &) const
-            {
-                return 0.0;
-            }
-    };
-
     /* P -> PP Processes */
 
     struct BToPiPi {
